Move Lifecyclehooks enum into compile_metadata.
This is to break a dependency between "compiler" and "core" angular.

PiperOrigin-RevId: 254988595
