+++ angular/lib/src/core/metadata/lifecycle_hooks.dart
@@ -7,16 +7,7 @@
 /// - `afterViewInit`,
 /// - `afterViewChecked`,
 /// - `onDestroy` (at the very end before destruction)
-enum LifecycleHooks {
-  onInit,
-  onDestroy,
-  doCheck,
-  afterChanges,
-  afterContentInit,
-  afterContentChecked,
-  afterViewInit,
-  afterViewChecked
-}
+library lifecycle_hooks;
 
 /// Implement this interface to get notified when any data-bound property of
 /// your directive is changed by the Angular framework.
