+++ angular/lib/src/source_gen/template_compiler/find_components.dart
@@ -17,7 +17,6 @@ import 'package:angular/src/compiler/view_compiler/property_binder.dart'
     show isPrimitiveTypeName;
 import 'package:angular/src/core/change_detection/constants.dart';
 import 'package:angular/src/core/metadata.dart';
-import 'package:angular/src/core/metadata/lifecycle_hooks.dart';
 import 'package:angular/src/source_gen/common/annotation_matcher.dart';
 import 'package:angular/src/source_gen/common/url_resolver.dart';
 import 'package:angular_compiler/angular_compiler.dart';
